Move everything into the main component
